requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configparser==7.2.0
cron_descriptor==2.0.6
Django==5.2
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
kombu==5.5.4
numpy==2.3.1
packaging==25.0
pandas==2.3.2
pika==1.3.2
pillow==11.2.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
unicode==2.9
vine==5.1.0
wcwidth==0.2.14